Fix matching of ES2015 unicode escapes
Example: `\u{123456}`.

I had mistakenly thought that only 1-6 hex digits were allowed between
the brackets, but _any_ number of hex digits is allowed according to the
ECMAScript lexical grammar. It is, however, an _early error_ if the
value of the hex digits is too large, so in practice you never end up
with more than 6 hex digits unless you have a large amount of leading
zeroes.